+++ myems-web/src/components/MyEMS/Space/SpaceOutput.js
@@ -33,6 +33,7 @@ import { comparisonTypeOptions } from '../common/ComparisonTypeOptions';
 import DateRangePickerWrapper from '../common/DateRangePickerWrapper';
 import { endOfDay } from 'date-fns';
 import AppContext from '../../../context/Context';
+import { Link } from 'react-router-dom';
 import blankPage from '../../../assets/img/generic/blank-page.png';
 
 const ChildSpacesTable = loadable(() => import('../common/ChildSpacesTable'));
@@ -80,8 +81,9 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
   const [basePeriodDateRange, setBasePeriodDateRange] = useState([
     current_moment
       .clone()
+      .subtract(7, 'days')
       .subtract(1, 'months')
-      .startOf('month')
+      .startOf('day')
       .toDate(),
     current_moment
       .clone()
@@ -92,7 +94,8 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
   const [reportingPeriodDateRange, setReportingPeriodDateRange] = useState([
     current_moment
       .clone()
-      .startOf('month')
+      .subtract(7, 'days')
+      .startOf('day')
       .toDate(),
     current_moment.toDate()
   ]);
@@ -186,8 +189,12 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
               .join('"label":')
           );
           setCascaderOptions(json);
+          // select root space name
           setSelectedSpaceName([json[0]].map(o => o.label));
+          // select root space ID
           setSelectedSpaceID([json[0]].map(o => o.value));
+          // load data with root space ID
+          loadData([json[0]].map(o => o.value));
         } else {
           toast.error(t(json.description));
         }
@@ -353,6 +360,10 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
   // Handler
   const handleSubmit = e => {
     e.preventDefault();
+    loadData(selectedSpaceID);
+  }
+
+  const loadData = (spaceID) => {
     // disable submit button
     setSubmitButtonDisabled(true);
     // show spinner
@@ -371,7 +382,7 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
       APIBaseURL +
       '/reports/spaceoutput?' +
       'spaceid=' +
-      selectedSpaceID +
+      spaceID +
       '&periodtype=' +
       periodType +
       '&baseperiodstartdatetime=' +
@@ -735,7 +746,9 @@ const SpaceOutput = ({ setRedirect, setRedirectUrl, t }) => {
       <div>
         <Breadcrumb>
           <BreadcrumbItem>{t('Space Data')}</BreadcrumbItem>
-          <BreadcrumbItem active>{t('Output')}</BreadcrumbItem>
+          <BreadcrumbItem active onClick={() => window.location.reload()}>
+            <Link to="/space/output">{t('Output')}</Link>
+          </BreadcrumbItem>
         </Breadcrumb>
       </div>
       <Card className="bg-light mb-3">