+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added control modes and control times
 ### Changed
 - replaced datetime.strftime() with datetime.isoformat() in myems-api
-- changed space energy category report to automaticly submit in myems-web
+- changed space reports to automaticly submit in myems-web
+- changed reporting period of space reports to one week in myems-web
 ### Fixed
 - fixed datetime isoformat issues in myems-api
 - fixed api key datetime issue in myems-api and myems-admin